Add test for calculation of a square when no moves are possible to it
In the case when a square cannot legally sum with any squares in the
column, `legalPrevColSums` would be an empty array. Math.max actually
returns negative infinity, which may not be a bad way to represent that
this square is dead and to ensure no squares in the next column would
sum with it.