Add an 'x' flag to SquareInfo to indicate that it's invalid/disabled
There are certain times when a square should be indicated as "disabled"
to make sure it is not considered as a candidate to be summed into the
next column:
1. A square has a value of 0, because the rules say the mining ends when
the miner reaches a square of 0.
2. A square is ineligible to be moved into, based on previous
eligibility restrictions.

In these cases, calculateNextColumn will now set a dirFlag of 'x' to
indicate the square is 'disabled'. When considering which square in the
previous column to sum with, calculateNextColumn will also ignore
squares with the 'x' flag.