Fix nasty bug in setting dirFlag logic
The SquareInfo object contains a 'dirFlags' field, which is a string
representing the immediate previous direction that was taken to
reach the current square.

When the SquareInfo of a particular square needs to be calculated
for a particular column, the three adjacent squares of the previous
column are examined for eligibility to be included in the calculation
of the cumulative sum. The dirFlags of each of these three adjacent
squares are examined in conjunction with their position to ensure that
including them in the sum would not mean that the same move is made
twice. For example, a square directly to the left of the
current square would be disqualified as a candidate to be summed
with the current square if it had a dirFlag of 'r', as that would mean
two moves to the right -  one that was used to reach the candidate,
and another move from the candidate to the current square.

Previously, the calculateNextColumn function did indeed disqualify such
candidates in the .filter function to ensure the maximum sum was not
calculated using these candidates. However, when it came time to set
the dirFlag, it was choosing the flag based solely on whether the
candidate's sum was the max sum. Therefore, previously disqualified
candidates were incorrectly re-qualified if their sum happened to match
the max sum calculated from qualified candidates.

This commit adds an additional check when setting the dirFlag to ensure
the position of the candidate and its dirflag make it eligible to be
summed into the current square.